pyproject.toml
[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "adh-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "faiss-cpu>=1.11.0",
    "mcp[cli]>=1.10.1",
    "nltk>=3.9.1",
    "openai-agents>=0.1.0",
    "patchright>=1.52.5",
    "pypdf>=5.7.0",
    "requests>=2.32.4",
    "tiktoken>=0.9.0",
]
